Progress

+++ src/main/java/com/Da_Technomancer/crossroads/API/packets/ISpinReceiver.java
@@ -2,6 +2,6 @@ package com.Da_Technomancer.crossroads.API.packets;
 
 public interface ISpinReceiver{
 	
-	public void receiveSpin(int identifier, float clientW, float angle);
+	public void receiveSpin(byte identifier, float clientW, float angle);
 	
 }
